Add root to the workspace but ignore subfolders (#4)
Add the root folder to the workspace, but ignore subfolders. This
enables us edit files at the root of the repo (such as the README.md
file), while keeping all the benefits of our multi-root workspace, and
without having the frontend and backend folders show up in the root
folder.