Fix all the text and simplify copy / paste of payload (#28)
- There was a regression and we were using the word envelope incorrectly
in a number of places.
- Tidy up the usage of those various terms, fix the label `for`
argument.
- The copy / paste buttons distinguish between URLs and payloads. This
is not that intuitive to a user. In general, we always just want to use
the URL, so now, payload refers to the full URL. And delete that old
payload button. Also, update Cipherly JS to support both the old payload
format (without the URL part). So if you just paste the part after the
hash, Cipherly will still decrypt correctly.